+++ WhistleWindLobotomyMod/Opponents/Ordeals/OrdealOpponent.cs
@@ -93,8 +93,8 @@ namespace WhistleWindLobotomyMod.Opponents {
 
             OrdealBannerManager.Instance.DisplayBanner(BattleSequencer.ordealType, true);
             this.SetSceneEffectsShown(true);
-            //AudioController.Instance.SetLoopAndPlay("first_warning", 1);
-            AudioController.Instance.SetLoopVolumeImmediate(0.3f, 1);
+            AudioController.Instance.SetLoopAndPlay("first_trumpet", 1);
+            AudioController.Instance.SetLoopVolumeImmediate(0.8f, 1);
             OrdealCounterManager.Instance.SetShown(true);
             yield return new WaitForSeconds(1.5f);